Some arch refactoring for scrutinizer happiness
* Removed stupid trait and added an abstract class
* Fixed some PHPDoc elements
* AbstractUser, Customer and AdminUser now return an array of Role instances
  to complain PHPDoc
