+++ src/Elcodi/Component/Core/Encoder/Interfaces/EncoderInterface.php
@@ -27,14 +27,14 @@ interface EncoderInterface
      *
      * @param mixed $data Data
      *
-     * @return string encoded data
+     * @return string|boolean encoded data
      */
     public function encode($data);
 
     /**
      * Decode incoming data
      *
-     * @param string $serializedData Serialized data
+     * @param string|boolean $serializedData Serialized data
      *
      * @return mixed Decoded data
      */
