+++ src/Elcodi/Component/EntityTranslator/Services/CachedEntityTranslationProvider.php
@@ -81,7 +81,7 @@ class CachedEntityTranslationProvider extends AbstractCacheWrapper implements En
      * @param string $entityField Field of entity
      * @param string $locale      Locale
      *
-     * @return string Value fetched
+     * @return string|boolean Value fetched
      */
     public function getTranslation(
         $entityType,
@@ -124,11 +124,11 @@ class CachedEntityTranslationProvider extends AbstractCacheWrapper implements En
     /**
      * Set translation
      *
-     * @param string $entityType       Type of entity
-     * @param string $entityId         Id of entity
-     * @param string $entityField      Field of entity
-     * @param string $translationValue Translated value
-     * @param string $locale           Locale
+     * @param string         $entityType       Type of entity
+     * @param string         $entityId         Id of entity
+     * @param string         $entityField      Field of entity
+     * @param string|boolean $translationValue Translated value
+     * @param string         $locale           Locale
      *
      * @return $this Self object
      */
