+++ src/Elcodi/Component/Plugin/Services/Traits/PluginUtilsTrait.php
@@ -26,7 +26,7 @@ trait PluginUtilsTrait
      * Load installed plugin bundles and return an array with them, indexed by
      * their namespaces
      *
-     * @return \Elcodi\Component\Plugin\Interfaces\PluginInterface[]|\Symfony\Component\HttpKernel\Bundle\Bundle[] Plugins installed
+     * @return \Symfony\Component\HttpKernel\Bundle\Bundle[] Plugins installed
      */
     protected function getInstalledPluginBundles(\Symfony\Component\HttpKernel\KernelInterface $kernel)
     {
@@ -34,11 +34,10 @@ trait PluginUtilsTrait
         $bundles = $kernel->getBundles();
 
         foreach ($bundles as $bundle) {
-
-            /**
-             * @var \Symfony\Component\HttpKernel\Bundle\Bundle|\Elcodi\Component\Plugin\Interfaces\PluginInterface $bundle
-             */
-            if ($bundle instanceof \Elcodi\Component\Plugin\Interfaces\PluginInterface) {
+            if (
+                $bundle instanceof \Symfony\Component\HttpKernel\Bundle\Bundle &&
+                $bundle instanceof \Elcodi\Component\Plugin\Interfaces\PluginInterface
+            ) {
                 $pluginNamespace = $bundle->getNamespace();
                 $plugins[$pluginNamespace] = $bundle;
             }
